Extract flare filler and rename test file
